+++ src/Actions/SendInvoiceAction.php
@@ -53,6 +53,10 @@ class SendInvoiceAction
      */
     public function handle(string $path, ?string $env = null, ?string $certPath = null, ?string $certPassword = null): InvoiceData
     {
+        if (! $this->storage->exists($path)) {
+            throw new \InvalidArgumentException("The XML file at path [{$path}] does not exist.");
+        }
+
         $signed = $this->storage->get($path);
 
         $object = new InvoiceXml($signed);